requirements.txt
asgiref==3.8.1
cloudinary==1.39.1
dj-database-url==2.1.0
dj-rest-auth==5.0.2
Django==5.0.3
django-allauth==0.61.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
oauthlib==3.2.2
pillow==10.2.0
psycopg2==2.9.9
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests-oauthlib==1.4.0
sqlparse==0.4.4
urllib3==2.2.1
